requirements.txt
playwright==1.44.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
gspread==6.1.2
google-auth==2.29.0
